+++ src/dune/dune_package.ml
@@ -362,7 +362,8 @@ module Or_meta = struct
          Dune_package package)
 
   let load p =
-    Vfile.load p ~f:(fun lang -> decode ~lang ~dir:(Path.parent_exn p))
+    let dir = Path.parent_exn p in
+    Vfile.load p ~f:(fun lang -> decode ~lang ~dir)
 
   let pp ~dune_version ppf t =
     let t = encode ~dune_version t in